tools/gui/requirements.txt
pygame>=2.5.0
python-chess>=1.9.0
pyperclip>=1.8.2
